requirements.txt
pytest==7.4
selenium==4.15
python-dotenv==1.0.0
allure-pytest==2.14
webdriver-manager==4.0.0
pytest-xdist==3.3.1  # For parallel test execution
